ai-service/ai_service/graph/state.py
from __future__ import annotations

from typing import Annotated, Any
from typing_extensions import TypedDict

from langgraph.graph.message import add_messages


class GraphState(TypedDict):
    messages: Annotated[list[Any], add_messages]
    pending_booking: dict | None
    account_id: str | None
