ai-service/pyproject.toml
[tool.poetry]
name = "nrg-flight-ai-service"
version = "0.1.0"
description = "FastAPI + LangGraph AI service for NRG Flight"
authors = ["NRG Flight <dev@example.com>"]
readme = "README.md"
packages = [{ include = "ai_service" }]

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }

# Config
pydantic-settings = "^2.6.1"

# Auth / Security (for decoding same JWT as api/)
python-jose = { version = "^3.3.0", extras = ["cryptography"] }

# DB (async)
SQLAlchemy = { version = "^2.0.36", extras = ["asyncio"] }
asyncpg = "^0.30.0"

# LangGraph
langchain-openai = "^0.2.0"
langchain-core = "^0.3.0"
langgraph = "^0.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[tool.ruff]
line-length = 100
